Generate in-toto build attestations in CD (wip)
Use in-toto cli to generate signed attestations for each build
target (sdist and wheel) and publish on the GH release page
along with the build artifacts.

*WARNING*
The signing key is an **unencrypted** ed25519 key generated with
`securesystemslib.interface.generate_and_write_unencrypted_ed25519_keypair`
stored as GitHub secret called `IN_TOTO_KEY`, and **written to a
file** in order for the in-toto CLI commands to access it.
Consider a more secure signing mechanism for non-POC use.

*TODO*
- Add comments to code in cd.yml

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>